randomizing agents and trials acc. to trialnr
